chore: remove useTheme
