+++ packages/docs/src/components/header/components/ColorDropdown.vue
@@ -20,14 +20,13 @@
 </template>
 
 <script lang="ts">
-import { useColors, useTheme } from 'vuestic-ui/src/main'
+import { useColors } from 'vuestic-ui/src/main'
 import { computed, defineComponent } from 'vue'
 
 export default defineComponent({
   name: 'DocsColorDropdown',
   setup () {
-    const { applyPreset, presets } = useTheme()
-    const { getColors, setColors } = useColors()
+    const { applyPreset, presets, getColors, setColors } = useColors()
     const capitalizeFirstLetter = (text: string) => text.charAt(0).toUpperCase() + text.slice(1)
 
     const colorsArray = computed(() => {
