+++ packages/docs/src/page-configs/services/colors-config/examples/css-variable.vue
@@ -25,7 +25,7 @@
 
 <script>
 import { computed } from 'vue'
-import { useColors, useTheme } from 'vuestic-ui/src/main'
+import { useColors } from 'vuestic-ui/src/main'
 
 export default {
   props: {
@@ -33,8 +33,7 @@ export default {
     buttonText: { type: String, default: 'Primary color button' },
   },
   setup () {
-    const { presets } = useTheme()
-    const { setColors, getColor } = useColors()
+    const { setColors, getColor, presets } = useColors()
 
     const colorsToChange = [
       presets.value.light.primary,
