+++ packages/ui/src/services/color-config/index.ts
@@ -1,2 +1,1 @@
 export * from './types'
-export * from './composables/useTheme'
